+++ tests/__init__.py
@@ -26,6 +26,10 @@ MUSICXML_IMPORT_EXPORT_TESTFILES = [
     os.path.join(MUSICXML_PATH, fn)
     for fn in ["test_note_ties.xml", "test_note_ties_divs.xml"]
 ]
+MUSICXML_SCORE_OBJECT_TESTFILES = [
+    os.path.join(MUSICXML_PATH, fn)
+    for fn in ["test_score_object.musicxml"]
+]
 MUSICXML_UNFOLD_TESTPAIRS = [
     (
         os.path.join(MUSICXML_PATH, fn1),
@@ -150,10 +154,6 @@ MUSICXML_PARTIAL_MEASURES_TESTFILES = [
 MEI_TESTFILES = [
     os.path.join(MEI_PATH, fn)
     for fn in [
-        "example_noMeasures_noBeams.mei",
-        "example_noMeasures_withBeams.mei",
-        "example_withMeasures_noBeams.mei",
-        "example_withMeasures_withBeams.mei",
         "Bach_Prelude.mei",
         "Schubert_An_die_Sonne_D.439.mei",
         "test_tuplets.mei",
@@ -171,6 +171,7 @@ MEI_TESTFILES = [
         "test_articulation.mei",
         "test_merge_voices2.mei",
         "CRIM_Mass_0030_4.mei",
+        "test_divs_tuplet.mei"
     ]
 ]
 
@@ -186,6 +187,11 @@ KERN_TESTFILES = [
         "variable_length_pr_bug.krn",
         "chor228.krn",
     ]
+]
+
+MUSESCORE_TESTFILES = [
+    os.path.join(DATA_PATH, "musescore", fn)
+    for fn in ["160.03_Pastorale.mscx"]
 ]
 
 KERN_TIES = [os.path.join(KERN_PATH, fn) for fn in ["tie_mismatch.krn"]]
@@ -234,4 +240,8 @@ TOKENIZER_TESTFILES = [
 
 MIDIEXPORT_TESTFILES = [
     os.path.join(DATA_PATH, "musicxml", "test_anacrusis.xml")
+]
+
+MIDIINPORT_TESTFILES = [
+    os.path.join(DATA_PATH, "midi", "bach_midi_score.mid")
 ]