+++ tests/test_load_score.py
@@ -37,16 +37,12 @@ class TestLoadScore(unittest.TestCase):
             + MATCH_IMPORT_EXPORT_TESTFILES
             + EXAMPLE_FILES
         ):
-            self.load_score(fn)
+            self.check_return_type(fn)
 
-    def load_score(self, fn):
-        try:
-            score = load_score(fn)
-            self.assertTrue(isinstance(score, Score))
-
-            for pp in score.part_structure:
-                self.assertTrue(type(pp) in (Part, PartGroup))
-            for pp in score.parts:
-                self.assertTrue(isinstance(pp, Part))
-        except NotSupportedFormatError:
-            self.assertTrue(False)
+    def check_return_type(self, fn):
+        score = load_score(fn)
+        self.assertTrue(isinstance(score, Score), f"results of load_score type are not Score for score {fn}.")
+        for pp in score.part_structure:
+            self.assertTrue(type(pp) in (Part, PartGroup), f"results of score.part_structure type are neither Part or PartGroup for score {fn}.")
+        for pp in score.parts:
+            self.assertTrue(isinstance(pp, Part), f"results of score.parts type are not Part for score {fn}.")